3_project_codebasics_q_and_a/chatbot_omdena/requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit>=1.22.0  # Updated to a version compatible with protobuf 4.25.1
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf==4.25.1  
torch  
InstructorEmbedding
sentence_transformers 
